+++ common/rmi/src/main/java/is/codion/common/rmi/server/SerializationWhitelist.java
@@ -20,7 +20,6 @@ import java.util.Collection;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.stream.Stream;
 
 import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.toList;
@@ -120,8 +119,23 @@ final class SerializationWhitelist {
       if (clazz == null) {
         return Status.ALLOWED;
       }
+      if (clazz.isArray()) {
+        return checkArrayInput(clazz);
+      }
 
       return checkInput(clazz.getName());
+    }
+
+    Status checkArrayInput(Class<?> arrayClass) {
+      Class<?> componentType = arrayClass.getComponentType();
+      while (componentType.isArray()) {
+        componentType = componentType.getComponentType();
+      }
+      if (componentType.isPrimitive()) {
+        return Status.ALLOWED;
+      }
+
+      return checkInput(componentType.getName());
     }
 
     Status checkInput(String classname) {
@@ -184,8 +198,8 @@ final class SerializationWhitelist {
     }
 
     private static Collection<String> readFileWhitelistItems(String whitelistFile) {
-      try (Stream<String> stream = Files.lines(Paths.get(whitelistFile))) {
-        return stream.collect(toSet());
+      try {
+        return new HashSet<>(Files.readAllLines(Paths.get(whitelistFile)));
       }
       catch (IOException e) {
         LOG.error("Unable to read serialization whitelist: " + whitelistFile);